Stamina works

+++ Assets/Scripts/HealthBar.cs
@@ -7,12 +7,12 @@ public class HealthBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetCurrentHealth(int currentHealth)
+    public void SetCurrentHealth(float currentHealth)
     {
         slider.value = currentHealth;
     }
 
-    public void SetMaxHealth(int maxHealth)
+    public void SetMaxHealth(float maxHealth)
     {
         slider.maxValue = maxHealth;
         slider.value = maxHealth;
