+++ Assets/Scripts/StaminaBar.cs
@@ -7,12 +7,12 @@ public class StaminaBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetCurrentStamina(int currentStamina)
+    public void SetCurrentStamina(float currentStamina)
     {
         slider.value = currentStamina;
     }
 
-    public void SetMaxStamina(int maxStamina)
+    public void SetMaxStamina(float maxStamina)
     {
         slider.maxValue = maxStamina;
         slider.value = maxStamina;
